Modularize LocationRepository
