+++ core/data/build.gradle.kts
@@ -25,6 +25,7 @@ plugins {
 android { namespace = "org.meshtastic.core.data" }
 
 dependencies {
+    implementation(projects.core.analytics)
     implementation(projects.core.database)
     implementation(projects.core.datastore)
     implementation(projects.core.di)
@@ -33,6 +34,7 @@ dependencies {
     implementation(projects.core.prefs)
     implementation(projects.core.proto)
 
+    implementation(libs.core.location.altitude)
     implementation(libs.kotlinx.coroutines.android)
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.timber)
